Test and deploy cofigs to solrcloud.
This commit replaces our travis deployment configurations with a new
circleci configuration.

In circleci we define one test workflow and one deployment workflow:
* test
* deploy

The deployment work process does the following:
* Push the configset to solrcloud (or fails the build)
* Creates a new collection based on the configset (or fails the build)
* Creates an alias to the collection if the build is non production (or fails the build)

Deployments are triggered by tag creation only:
* Non production deployments happen when a tag is pushed that ends with -qa, -stage, or -rc.
* All other tags are assumed to be production.

Alias are defined as follows.
* The production alias is always tul_cob-az
* Non production aliases are tul_cob-az($ALIAS_SUFFIX)

The $ALIAS_SUFFIX is generated from the tag and it's defined as text that
meets the following regex criteria:

```
/.*(-[A-z0-9]+)$/\1/
``