Test and deploy cofigs to solrcloud.
This commit replaces our travis deployment configurations with a new
circleci configuration.

In circleci we define one test workflow and one deployment workflow:
* test
* deploy

The deployment work process does the following:
* Push the configset to solrcloud (or fails the build)
* Creates a new collection based on the configset (or fails the build)
* Creates an alias to the collection if alias does not already exist (or
fails build)

Deployments are triggered by tag creation only:

Alias are defined as follows.
* tul_cob-az($ALIAS_SUFFIX)

The $ALIAS_SUFFIX is generated from the tag and it's defined as text that
meets the following regex criteria:

```
/.*(-[A-z0-9]+)$/\1/
``